update:修改数据库查询部分
