+++ app/build.gradle
@@ -84,8 +84,9 @@ dependencies {
 
     implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0-alpha04'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.0.0'
-//    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.1.1'
-//    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.1.1'
+    implementation "org.jetbrains.anko:anko:0.10.8"
+    implementation "org.jetbrains.anko:anko-design:0.10.8" // For SnackBars
+    implementation "org.jetbrains.anko:anko-coroutines:0.10.8"
     implementation files('libs/TalkingdataAnalytics.jar')
     // 打google play包用下面这个
 //    implementation files('libs/GooglePlayTalkingDataAppAnalytics.jar')
