+++ app/src/main/java/info/free/scp/SCPConstants.kt
@@ -7,7 +7,7 @@ package info.free.scp
 
 object SCPConstants {
     const val PACKAGE_NAME = "info.free.scp"
-    const val DB_NAME = "scp_info.db"
+    const val DB_NAME = "scp_data.db"
     // category type
     object Entry {
         const val SCP_DOC = 0 // SCP文档，包含其他文档
