+++ app/src/main/java/info/free/scp/db/AppDatabase.kt
@@ -5,12 +5,14 @@ import androidx.room.RoomDatabase
 import androidx.room.Database
 import androidx.room.Room
 import info.free.scp.SCPConstants.DB_NAME
+import info.free.scp.bean.ScpDetail
 import info.free.scp.bean.ScpModel
 
 
-@Database(entities = [ScpModel::class], version = 1)
+@Database(entities = [ScpModel::class, ScpDetail::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
-    abstract fun scpDao(): ArticleDao
+    abstract fun scpDao(): ScpDao
+    abstract fun detailDao(): DetailDao
 
     companion object {
         private var INSTANCE: AppDatabase? = null
