+++ app/src/main/java/info/free/scp/view/ScpRemoteView.kt
@@ -5,7 +5,7 @@ import android.content.Intent
 import android.widget.RemoteViews
 import info.free.scp.R
 import info.free.scp.ScpApplication
-import info.free.scp.db.ScpDao
+import info.free.scp.db.ScpDataHelper
 import info.free.scp.view.detail.DetailActivity
 import java.util.*
 import java.util.Calendar.DAY_OF_WEEK
@@ -17,7 +17,7 @@ class ScpRemoteView(packageName: String) : RemoteViews(packageName, R.layout.lay
         setTextViewText(R.id.tv_widget_date, "${getWeekDayString(today.get(DAY_OF_WEEK))} " +
                 "${today.get(Calendar.YEAR)}.${today.get(Calendar.MONTH) + 1}." +
                 "${today.get(Calendar.DAY_OF_MONTH)}")
-        val todayScp = ScpDao.getInstance().getRandomScp()
+        val todayScp = ScpDataHelper.getInstance().getRandomScp()
         todayScp?.let {
             setTextViewText(R.id.tv_widget_scp_title, it.title)
             val intent = Intent(ScpApplication.context, DetailActivity::class.java)
