+++ app/src/main/java/info/free/scp/view/category/ScpListFragment.kt
@@ -2,7 +2,6 @@ package info.free.scp.view.category
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -10,7 +9,7 @@ import android.view.ViewGroup
 import info.free.scp.R
 import info.free.scp.SCPConstants
 import info.free.scp.bean.ScpModel
-import info.free.scp.db.ScpDao
+import info.free.scp.db.ScpDataHelper
 import info.free.scp.util.PreferenceUtil
 import info.free.scp.util.Toaster
 import info.free.scp.view.base.BaseAdapter
@@ -78,80 +77,80 @@ class ScpListFragment : BaseFragment() {
                 // 0,499,999
                 val start = if (clickPosition == 0) 0 else clickPosition * categoryCount
                 val limit = categoryCount
-                scpList?.addAll(ScpDao.getInstance().getScpByTypeAndRange(SCPConstants.ScpType.SAVE_SERIES, start, limit))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByTypeAndRange(SCPConstants.ScpType.SAVE_SERIES, start, limit))
             }
             SCPConstants.Category.SERIES_CN -> {
                 val start = if (clickPosition == 0) 0 else clickPosition * categoryCount
                 val limit = categoryCount
-                scpList?.addAll(ScpDao.getInstance().getScpByTypeAndRange(SCPConstants.ScpType.SAVE_SERIES_CN, start, limit))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByTypeAndRange(SCPConstants.ScpType.SAVE_SERIES_CN, start, limit))
 
             }
             SCPConstants.Category.SCP_EX -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX))
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX_CN))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX_CN))
             }
 //            SCPConstants.Category.SCP_EX_CN -> {
-//                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX_CN))
+//                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_EX_CN))
 //            }
             SCPConstants.Category.SCP_ABNORMAL -> {
-                scpList?.addAll(ScpDao.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_ABNORMAL))
+                scpList?.addAll(ScpDataHelper.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_ABNORMAL))
                 // 三句话外围
-                scpList?.add(ScpDao.getInstance().getOneScpModelByLink("/short-stories"))
+                scpList?.add(ScpDataHelper.getInstance().getOneScpModelByLink("/short-stories"))
             }
             SCPConstants.Category.ABOUT_INFO -> {
                 // 相关材料
-                scpList?.addAll(ScpDao.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_INFO))
+                scpList?.addAll(ScpDataHelper.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_INFO))
             }
             SCPConstants.Category.ABOUT_INTRO -> {
                 // 相关材料
-                scpList?.addAll(ScpDao.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_INTRO))
+                scpList?.addAll(ScpDataHelper.getInstance().getSinglePageByType(SCPConstants.ScpType.SAVE_INTRO))
             }
 
             SCPConstants.Category.SCP_ARCHIVES -> {
                 when (clickPosition) {
-                    0->scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_ARCHIVED))
-                    1->scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_DECOMMISSIONED))
-                    2->scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_REMOVED))
+                    0->scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_ARCHIVED))
+                    1->scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_DECOMMISSIONED))
+                    2->scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_REMOVED))
                 }
             }
             SCPConstants.Category.TALES -> {
-                scpList?.addAll(ScpDao.getInstance().getTaleByTypeAndLetter(SCPConstants.ScpType.SAVE_TALES
+                scpList?.addAll(ScpDataHelper.getInstance().getTaleByTypeAndLetter(SCPConstants.ScpType.SAVE_TALES
                         , taleCategory[clickPosition]))
             }
             SCPConstants.Category.TALES_CN -> {
-                scpList?.addAll(ScpDao.getInstance().getTaleByTypeAndLetter(SCPConstants.ScpType.SAVE_TALES_CN
+                scpList?.addAll(ScpDataHelper.getInstance().getTaleByTypeAndLetter(SCPConstants.ScpType.SAVE_TALES_CN
                         ,taleCategory[clickPosition]))
             }
             SCPConstants.Category.STORY_SERIES -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_STORY_SERIES))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_STORY_SERIES))
             }
             SCPConstants.Category.STORY_SERIES_CN -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_STORY_SERIES_CN))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_STORY_SERIES_CN))
             }
             SCPConstants.Category.JOKE -> {
                 // 内容较少，直接全部加载
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_JOKE))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_JOKE))
             }
             SCPConstants.Category.JOKE_CN -> {
                 // 内容较少，直接全部加载
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_JOKE_CN))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_JOKE_CN))
             }
             SCPConstants.Category.SETTINGS -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_SETTINGS))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_SETTINGS))
             }
             SCPConstants.Category.SETTINGS_CN -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_SETTINGS_CN))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_SETTINGS_CN))
             }
             SCPConstants.Category.CONTEST -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_CONTEST))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_CONTEST))
             }
             SCPConstants.Category.CONTEST_CN -> {
-                scpList?.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_CONTEST_CN))
+                scpList?.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_CONTEST_CN))
             }
 
             SCPConstants.Category.TALES_BY_TIME -> {
                 if (taleTimeList.isEmpty()) {
-                    taleTimeList.addAll(ScpDao.getInstance().getScpByType(SCPConstants.ScpType.SAVE_TALES_CN))
+                    taleTimeList.addAll(ScpDataHelper.getInstance().getScpByType(SCPConstants.ScpType.SAVE_TALES_CN))
                 }
                 when (clickPosition) {
                     0 -> {
