+++ app/src/main/java/info/free/scp/view/search/SearchHolder.kt
@@ -1,17 +1,13 @@
 package info.free.scp.view.search
 
 import android.content.Context
-import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.View.VISIBLE
-import info.free.scp.R
 import info.free.scp.SCPConstants.LATER_TYPE
-import info.free.scp.bean.ScpModel
 import info.free.scp.bean.SimpleScp
-import info.free.scp.db.ScpDao
+import info.free.scp.db.ScpDataHelper
 import info.free.scp.util.ThemeUtil
 import info.free.scp.util.Toaster
-import info.free.scp.util.Utils
 import kotlinx.android.synthetic.main.item_search.view.*
 
 /**
@@ -41,11 +37,11 @@ class SearchHolder(view: View) : androidx.recyclerview.widget.RecyclerView.ViewH
             }
             itemView.btn_search_read_later.setOnClickListener {
                 if (isInLaterViewList) {
-                    ScpDao.getInstance().deleteViewListItem(link, LATER_TYPE)
+                    ScpDataHelper.getInstance().deleteViewListItem(link, LATER_TYPE)
                     isInLaterViewList = false
                     it.setBackgroundColor(ThemeUtil.unClickBtn)
                 } else {
-                    ScpDao.getInstance().insertViewListItem(link, title, LATER_TYPE)
+                    ScpDataHelper.getInstance().insertViewListItem(link, title, LATER_TYPE)
                     Toaster.show("已加入待读列表")
                     isInLaterViewList = true
                     it.setBackgroundColor(ThemeUtil.clickedBtn)
